Move avatar editing to user profile page
