+++ app/static/js/main.js
@@ -18,6 +18,7 @@ import "./traces/_preview.js"
 import "./traces/_upload.js"
 import "./user/_account-confirm.js"
 import "./user/_login.js"
+import "./user/_profile.js"
 import "./user/_reset-password.js"
 import "./user/_signup.js"
 import "./user/_terms.js"
