+++ app/static/js/settings/_index.js
@@ -17,13 +17,6 @@ if (settingsBody) {
     const settingsForm = settingsBody.querySelector("form.settings-form")
     const displayNameInput = settingsForm.elements.display_name
     const displayNameBlacklist = displayNameInput.dataset.blacklist
-    const avatarForm = settingsBody.querySelector("form.avatar-form")
-    const avatarTypeInput = avatarForm.elements.avatar_type
-    const avatarFileInput = avatarForm.elements.avatar_file
-    const uploadAvatarButton = avatarForm.querySelector("button.upload-avatar")
-    const useGravatarButton = avatarForm.querySelector("button.use-gravatar")
-    const removeAvatarButton = avatarForm.querySelector("button.remove-avatar")
-    const avatars = document.querySelectorAll("img.avatar")
 
     const onSettingsFormSuccess = () => {
         location.reload()
@@ -44,38 +37,5 @@ if (settingsBody) {
         return result
     }
 
-    const onAvatarFileChange = () => {
-        avatarTypeInput.value = "custom"
-        avatarForm.requestSubmit()
-    }
-
-    const onUploadAvatarClick = () => {
-        avatarFileInput.click()
-    }
-
-    const onUseGravatarClick = () => {
-        avatarTypeInput.value = "gravatar"
-        avatarFileInput.value = ""
-        avatarForm.requestSubmit()
-    }
-
-    const onRemoveAvatarClick = () => {
-        avatarTypeInput.value = "default"
-        avatarFileInput.value = ""
-        avatarForm.requestSubmit()
-    }
-
-    // On successful avatar upload, update avatar images
-    const onAvatarFormSuccess = (data) => {
-        for (const avatar of avatars) avatar.src = data.avatar_url
-    }
-
     configureStandardForm(settingsForm, onSettingsFormSuccess, onSettingsClientValidation)
-    configureStandardForm(avatarForm, onAvatarFormSuccess)
-
-    // Listen for events
-    avatarFileInput.addEventListener("change", onAvatarFileChange)
-    uploadAvatarButton.addEventListener("click", onUploadAvatarClick)
-    useGravatarButton.addEventListener("click", onUseGravatarClick)
-    removeAvatarButton.addEventListener("click", onRemoveAvatarClick)
 }
